Update UserField and AvailableField
Added method CheckNearbyShip, which checks nearby cells for the presence of ship,and he allows user to place ship there or no. Added method CheckPartition, which checks nearby cell on all directions, and allow user to add cell to the current ship or no. Added method CheckAllShip, which finds activated cells, so ships, and he paints ship with color according to the size of his. Showing available ships to choose for user is being on the procces.